Support building libddwaf from source (#31)
Add source-static and source-shared features backed by the published
libddwaf-src 2.1.0 crate. Validate incompatible feature combinations and
select the headers and libraries exported by the source crate without
downloading binary release artifacts.

Support source-built libraries in both embedded and linker-driven
dynamic modes while retaining the required macOS libc++ linkage. Remove
the development-only link-stdcxx feature and document how to pass custom
C++ runtime linker arguments directly when needed.

Document the new build modes and skip release-version assertions when a
source build is selected. Add CI coverage for static, embedded shared,
and dynamic-link variants, including verification of runtime linkage.

See also: https://github.com/DataDog/dd-trace-php/issues/4174

+++ crates/libddwaf-sys/tests/tests.rs
@@ -436,7 +436,7 @@ fn test_compact_and_large_containers_are_equivalent() {
         },
     };
     let mut large_array_data = _ddwaf_object_large_array::default();
-    large_array_data.set__type(u64::from(DDWAF_OBJ_LARGE_ARRAY));
+    large_array_data.set__type(u64::from(DDWAF_OBJ_LARGE_ARRAY as u8));
     large_array_data.set_size(1);
     large_array_data.set_capacity(1);
     large_array_data.ptr = array_items.as_mut_ptr();
@@ -473,7 +473,7 @@ fn test_compact_and_large_containers_are_equivalent() {
         },
     };
     let mut large_map_data = _ddwaf_object_large_map::default();
-    large_map_data.set__type(u64::from(DDWAF_OBJ_LARGE_MAP));
+    large_map_data.set__type(u64::from(DDWAF_OBJ_LARGE_MAP as u8));
     large_map_data.set_size(1);
     large_map_data.set_capacity(1);
     large_map_data.ptr = map_items.as_mut_ptr();
